Refactoring ContractDeployFragment and ProfileFragment, bugfixing, persist profiles for account

+++ main/java/ch/uzh/ifi/csg/contract/contract/IPurchaseContract.java
@@ -1,5 +1,7 @@
 package ch.uzh.ifi.csg.contract.contract;
 
+import java.math.BigInteger;
+
 import ch.uzh.ifi.csg.contract.async.promise.SimplePromise;
 import ch.uzh.ifi.csg.contract.event.IContractObservable;
 import ch.uzh.ifi.csg.contract.service.account.UserProfile;
@@ -9,9 +11,10 @@ public interface IPurchaseContract extends IContractObservable {
 	String getContractAddress();
     UserProfile getUserProfile();
     void setUserProfile(UserProfile profile);
+
     SimplePromise<String> seller();
     SimplePromise<String> abort();
-    SimplePromise<Integer> value();
+    SimplePromise<BigInteger> value();
     SimplePromise<String> title();
     SimplePromise<String> description();
     SimplePromise<String> buyer();
