+++ main/java/ch/uzh/ifi/csg/contract/service/account/AccountService.java
@@ -1,6 +1,7 @@
 package ch.uzh.ifi.csg.contract.service.account;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 import ch.uzh.ifi.csg.contract.async.promise.SimplePromise;
@@ -14,7 +15,7 @@ public interface AccountService {
     SimplePromise<List<Account>> getAccounts();
     SimplePromise<Account> createAccount(String alias, String password);
     SimplePromise<Boolean> unlockAccount(Account account, String password);
-    SimplePromise<BigDecimal> getAccountBalance(String account);
+    SimplePromise<BigInteger> getAccountBalance(String account);
     UserProfile getAccountProfile(String account);
     void saveAccountProfile(String accountId, UserProfile profile);
 }
